Request de productos por comercio

+++ src/main/java/com/portal/comercio/Models/comercios.java
@@ -4,13 +4,10 @@ import java.io.Serializable;
 import java.sql.Date;
 
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.Column;
 
 @Entity
@@ -26,7 +23,7 @@ public class Comercios implements Serializable {
     @Getter
     @Setter
     @Column(name = "id_comercio")
-    private int idComercio;
+    private Long idComercio;
 
     @Getter
     @Setter
